requirements.txt
python-multipart==0.0.20
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
psycopg[binary]==3.3.3
PyJWT>=2.8.0